Move load/import setup to its own module.
